Move output encoding utilities into core
