+++ platform/Win/MSX1PaletteQuantizer_CLI.vcxproj
@@ -154,12 +154,14 @@
   <ItemGroup>
     <ClInclude Include="..\..\src\cli\lodepng.h" />
     <ClInclude Include="..\..\src\core\MSX1PQCore.h" />
+    <ClInclude Include="..\..\src\core\MSX1PQOutput.h" />
     <ClInclude Include="..\..\src\core\MSX1PQPalettes.h" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="..\..\src\cli\lodepng.cpp" />
     <ClCompile Include="..\..\src\cli\msx1pq_cli.cpp" />
     <ClCompile Include="..\..\src\core\MSX1PQCore.cpp" />
+    <ClCompile Include="..\..\src\core\MSX1PQOutput.cpp" />
     <ClCompile Include="..\..\src\core\MSX1PQPalettes.cpp" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
